+++ wn/__init__.py
@@ -26,7 +26,7 @@ __all__ = (
 from wn._meta import __version__
 from wn._exceptions import Error
 from wn._config import config  # noqa: F401
-from wn._db import add, remove
+from wn._db import add, remove, is_schema_compatible
 from wn._download import download
 from wn._core import (
     lexicons, Lexicon,
@@ -35,3 +35,12 @@ from wn._core import (
     synset, synsets, Synset,
     Wordnet
 )
+
+if not is_schema_compatible():
+    installed = '\n  '.join(f'{lex.id}:{lex.version}' for lex in lexicons())
+    raise Error(
+        "Wn's schema has changed and is no longer compatible with the "
+        f"database. Please move or delete {config.database_path} and "
+        "rebuild it. Note that the following lexicons are currently "
+        f"installed:\n  {installed}"
+    )